Fix failure to re-authenticate after session loss
Logging into the IntesisHome web UI via a browser, or power cycling the
AC unit, both invalidate the existing session ID used by the library.

The code attempting to handle this was simply checking for the existence
of a "success" key in the JSON response - however it didn't check if
this value was false, so the case where `success=false` and `error` was
set was not handled.

This changeset checks that the success value is "truthy", defaulting to
false if the key is not found. It then proceeds to more comprehensive
error handling coverage.

This has been validated to operate correctly on a Fujitsu
FJ-AC-WIFI-1B module.

Additionally, logging of the IP address of the AC unit has been added to
the "Sending intesishome_local command ..." message, along with all the
potential error paths. This is particularly useful when controlling
multiple AC units.